src/main/java/com/inventory/domain/Product.java
/**
 * 
 */
package com.inventory.domain;

import java.io.Serializable;

/**
 * @author Dinesh Sapkota
 * 10 Jul 2017
 */
public class Product implements Serializable {

    private String description;
    private Double price;
    
    public String getDescription() {
        return description;
    }
    
    public void setDescription(String description) {
        this.description = description;
    }
    
    public Double getPrice() {
        return price;
    }
    
    public void setPrice(Double price) {
        this.price = price;
    }
    
    public String toString() {
        StringBuffer buffer = new StringBuffer();
        buffer.append("Description: " + description + ";");
        buffer.append("Price: " + price);
        return buffer.toString();
    }
}
